7장-연습 . init - columnParser 2

+++ src/test/java/cahhp07/practice/ColumnsIndexInfo.java
@@ -2,14 +2,14 @@ package cahhp07.practice;
 
 public class ColumnsIndexInfo {
 //    String defaultColumns = "Dimension.DATE,Ad unit 1,Ad unit 2,Ad unit ID 1,Ad unit ID 2,Column.TOTAL_AD_REQUESTS,Column.TOTAL_LINE_ITEM_LEVEL_IMPRESSIONS,Column.TOTAL_LINE_ITEM_LEVEL_CLICKS,Column.TOTAL_LINE_ITEM_LEVEL_CPM_AND_CPC_REVENUE,Column.TOTAL_ACTIVE_VIEW_VIEWABLE_IMPRESSIONS_RATE";
-    private short dateIndex;
-    private short unitNameIndex;
-    private short unitIdIndex;
-    private short requestIndex;
-    private short impressionIndex;
-    private short clickIndex;
-    private short revenueIndex;
-    private short viewablityIndex;
+    private byte dateIndex;
+    private byte unitNameIndex;
+    private byte unitIdIndex;
+    private byte requestIndex;
+    private byte impressionIndex;
+    private byte clickIndex;
+    private byte revenueIndex;
+    private byte viewablityIndex;
 
     @Override
     public String toString() {
@@ -25,67 +25,67 @@ public class ColumnsIndexInfo {
                 '}';
     }
 
-    public short getDateIndex() {
+    public byte getDateIndex() {
         return dateIndex;
     }
 
-    public void setDateIndex(short dateIndex) {
+    public void setDateIndex(byte dateIndex) {
         this.dateIndex = dateIndex;
     }
 
-    public short getUnitNameIndex() {
+    public byte getUnitNameIndex() {
         return unitNameIndex;
     }
 
-    public void setUnitNameIndex(short unitNameIndex) {
+    public void setUnitNameIndex(byte unitNameIndex) {
         this.unitNameIndex = unitNameIndex;
     }
 
-    public short getUnitIdIndex() {
+    public byte getUnitIdIndex() {
         return unitIdIndex;
     }
 
-    public void setUnitIdIndex(short unitIdIndex) {
+    public void setUnitIdIndex(byte unitIdIndex) {
         this.unitIdIndex = unitIdIndex;
     }
 
-    public short getRequestIndex() {
+    public byte getRequestIndex() {
         return requestIndex;
     }
 
-    public void setRequestIndex(short requestIndex) {
+    public void setRequestIndex(byte requestIndex) {
         this.requestIndex = requestIndex;
     }
 
-    public short getImpressionIndex() {
+    public byte getImpressionIndex() {
         return impressionIndex;
     }
 
-    public void setImpressionIndex(short impressionIndex) {
+    public void setImpressionIndex(byte impressionIndex) {
         this.impressionIndex = impressionIndex;
     }
 
-    public short getClickIndex() {
+    public byte getClickIndex() {
         return clickIndex;
     }
 
-    public void setClickIndex(short clickIndex) {
+    public void setClickIndex(byte clickIndex) {
         this.clickIndex = clickIndex;
     }
 
-    public short getRevenueIndex() {
+    public byte getRevenueIndex() {
         return revenueIndex;
     }
 
-    public void setRevenueIndex(short revenueIndex) {
+    public void setRevenueIndex(byte revenueIndex) {
         this.revenueIndex = revenueIndex;
     }
 
-    public short getViewablityIndex() {
+    public byte getViewablityIndex() {
         return viewablityIndex;
     }
 
-    public void setViewablityIndex(short viewablityIndex) {
+    public void setViewablityIndex(byte viewablityIndex) {
         this.viewablityIndex = viewablityIndex;
     }
 }
